+++ src/Bridge/ManticoreSearch/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.11
+----
+
+ * Normalize the `endpoint` URL and tolerate a trailing slash
+ * [BC BREAK] Rename the `$host` constructor argument to `$endpoint`
+
 0.1
 ---