+++ src/Bridge/Supabase/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.11
+----
+
+ * Normalize the `endpoint` URL and tolerate a trailing slash
+ * [BC BREAK] Rename the `$url` constructor/factory argument to `$endpoint`
+
 0.7
 ---