+++ src/Bridge/Typesense/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.11
 ----
 
+ * Normalize the `endpoint` URL and tolerate a missing trailing slash on path-prefixed endpoints
  * Introduce a `StoreFactory`
  * [BC BREAK] Add support for `ScopingHttpClient` in `Store`
  * [BC BREAK] The `endpointUrl` parameter for `Store` has been removed